FPLA-1436: Refactored returnApplication defintions to use caseEventToComplexType
