+++ ccd-definition/AuthorisationCaseField/court-admin.json
@@ -831,12 +831,5 @@
     "CaseFieldID": "returnApplication",
     "UserRole": "caseworker-publiclaw-courtadmin",
     "CRUD": "CRUD"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "returnedDocumentBundle",
-    "UserRole": "caseworker-publiclaw-courtadmin",
-    "CRUD": "CRUD"
   }
 ]
