+++ ccd-definition/AuthorisationCaseField/la-solicitor.json
@@ -800,12 +800,5 @@
     "CaseFieldID": "returnApplication",
     "UserRole": "[LASOLICITOR]",
     "CRUD": "R"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "returnedDocumentBundle",
-    "UserRole": "[LASOLICITOR]",
-    "CRUD": "R"
   }
 ]
