+++ ccd-definition/CaseEventToComplexTypes/returnApplication/ReturnApplication.json
@@ -1,0 +1,50 @@
+[
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "ReturnApplication",
+    "CaseEventID": "returnApplication",
+    "CaseFieldID": "returnApplication",
+    "ListElementCode": "reason",
+    "FieldDisplayOrder": 1,
+    "DisplayContext": "MANDATORY"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "ReturnApplication",
+    "CaseEventID": "returnApplication",
+    "CaseFieldID": "returnApplication",
+    "ListElementCode": "note",
+    "FieldDisplayOrder": 2,
+    "DisplayContext": "MANDATORY"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "ReturnApplication",
+    "CaseEventID": "returnApplication",
+    "CaseFieldID": "returnApplication",
+    "ListElementCode": "submittedDate",
+    "FieldDisplayOrder": 3,
+    "DisplayContext": "READONLY",
+    "FieldShowCondition": "note = \"DO NOT SHOW\""
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "ReturnApplication",
+    "CaseEventID": "returnApplication",
+    "CaseFieldID": "returnApplication",
+    "ListElementCode": "returnedDate",
+    "FieldDisplayOrder": 4,
+    "DisplayContext": "READONLY",
+    "FieldShowCondition": "note = \"DO NOT SHOW\""
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "ReturnApplication",
+    "CaseEventID": "returnApplication",
+    "CaseFieldID": "returnApplication",
+    "ListElementCode": "document",
+    "FieldDisplayOrder": 5,
+    "DisplayContext": "READONLY",
+    "FieldShowCondition": "note = \"DO NOT SHOW\""
+  }
+]
