+++ ccd-definition/CaseEventToFields/returnApplication.json
@@ -5,21 +5,9 @@
     "CaseEventID": "returnApplication",
     "CaseFieldID": "returnApplication",
     "PageFieldDisplayOrder": 1,
-    "DisplayContext": "MANDATORY",
+    "DisplayContext": "COMPLEX",
     "PageID": 1,
     "PageDisplayOrder": 1,
-    "PageColumnNumber": 1
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "returnApplication",
-    "CaseFieldID": "returnedDocumentBundle",
-    "PageFieldDisplayOrder": 2,
-    "DisplayContext": "READONLY",
-    "PageID": 1,
-    "PageDisplayOrder": 1,
-    "FieldShowCondition" : "returnApplication=\"DO NOT SHOW\"",
     "PageColumnNumber": 1
   }
 ]
