+++ ccd-definition/CaseTypeTab/14_Overview.json
@@ -16,16 +16,6 @@
     "TabID": "OverviewTab",
     "TabLabel": "Overview",
     "TabDisplayOrder": 14,
-    "CaseFieldID": "returnedDocumentBundle",
-    "TabFieldDisplayOrder": 1
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "Channel": "CaseWorker",
-    "TabID": "OverviewTab",
-    "TabLabel": "Overview",
-    "TabDisplayOrder": 14,
     "CaseFieldID": "caseCompletionDate",
     "TabFieldDisplayOrder": 2
   },
