+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/NotifyGatekeeperControllerAboutToStartTest.java
@@ -8,10 +8,9 @@ import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
+import uk.gov.hmcts.reform.fpl.model.ReturnApplication;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
 import uk.gov.hmcts.reform.fpl.model.common.EmailAddress;
-import uk.gov.hmcts.reform.fpl.model.returnapplication.ReturnApplication;
-import uk.gov.hmcts.reform.fpl.model.returnapplication.ReturnedDocumentBundle;
 import uk.gov.hmcts.reform.fpl.service.ValidateGroupService;
 
 import java.util.List;
@@ -21,7 +20,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static uk.gov.hmcts.reform.fpl.controllers.ReturnApplicationController.RETURNED_DOCUMENT_BUNDLE;
 import static uk.gov.hmcts.reform.fpl.controllers.ReturnApplicationController.RETURN_APPLICATION;
 import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.element;
 import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.wrapElements;
@@ -75,7 +73,6 @@ class NotifyGatekeeperControllerAboutToStartTest extends AbstractControllerTest 
             VALID_FAMILY_MAN_NUMBER, SUBMITTED));
 
         assertThat(callbackResponse.getData().get(RETURN_APPLICATION)).isNull();
-        assertThat(callbackResponse.getData().get(RETURNED_DOCUMENT_BUNDLE)).isNull();
     }
 
     private CaseDetails caseDetails(String familyManNumber, String state) {
@@ -84,9 +81,6 @@ class NotifyGatekeeperControllerAboutToStartTest extends AbstractControllerTest 
             .data(Map.of(
                     RETURN_APPLICATION, ReturnApplication.builder()
                         .note("Reason")
-                        .build(),
-                    RETURNED_DOCUMENT_BUNDLE, ReturnedDocumentBundle.builder()
-                        .returnedDate("Returned date")
                         .build(),
                 "familyManCaseNumber", familyManNumber,
                 "gatekeeperEmails", wrapElements(
