+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/CaseData.java
@@ -28,8 +28,6 @@ import uk.gov.hmcts.reform.fpl.model.order.generated.FurtherDirections;
 import uk.gov.hmcts.reform.fpl.model.order.generated.GeneratedOrder;
 import uk.gov.hmcts.reform.fpl.model.order.generated.InterimEndDate;
 import uk.gov.hmcts.reform.fpl.model.order.selector.ChildSelector;
-import uk.gov.hmcts.reform.fpl.model.returnapplication.ReturnApplication;
-import uk.gov.hmcts.reform.fpl.model.returnapplication.ReturnedDocumentBundle;
 import uk.gov.hmcts.reform.fpl.utils.ElementUtils;
 import uk.gov.hmcts.reform.fpl.validation.groups.CaseExtensionGroup;
 import uk.gov.hmcts.reform.fpl.validation.groups.DateOfIssueGroup;
@@ -440,5 +438,4 @@ public class CaseData {
     private final CaseExtensionTime caseExtensionTimeList;
     private final CaseExtensionTime caseExtensionTimeConfirmationList;
     private final ReturnApplication returnApplication;
-    private final ReturnedDocumentBundle returnedDocumentBundle;
 }
