--- service/src/main/java/uk/gov/hmcts/reform/fpl/model/returnapplication/ReturnApplication.java
@@ -1,16 +1,0 @@
-package uk.gov.hmcts.reform.fpl.model.returnapplication;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import uk.gov.hmcts.reform.fpl.enums.ReturnedApplicationReasons;
-
-import java.util.List;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class ReturnApplication {
-    private final List<ReturnedApplicationReasons> reason;
-    private final String note;
-}
